Add looser warnings
This adds looser warnings, better suited for the new count of 7
used algorithms.  This means that less sentences will be marked as
difficult.

Additionally, for an algorithm to be counted, it now has to have
a result higher than the target age, where it previously was higher
or equal to the target age.

This also fixes the tests for the new algorithm.